FEATURE split vagrant setup
Setting up the testenv is now a two-step process.
First, the required boxes are built from the Vagrantfile in /boxes.
The boxes are then used in the main Vagrantfile.
This reduces overhead when rebuilding the setup, since we now start with pre-configure boxes which are already set up for salt.